+++ test/use_case/food/FoodInteractorClassTest.java
@@ -22,9 +22,9 @@ public class FoodInteractorClassTest {
         // Create a stub of FoodDataAccessInterface with predefined responses
         stubDataAccess = new FoodDataAccessInterface() {
             @Override
-            public HashMap<String, Double> getFoodNutritionalValues(String foodName, Float quantity) {
-                HashMap<String, Double> nutritionalValues = new HashMap<>();
-                nutritionalValues.put("Fiber", 4.0);
+            public HashMap<String, Float> getFoodNutritionalValues(String foodName, Float quantity) {
+                HashMap<String, Float> nutritionalValues = new HashMap<>();
+                nutritionalValues.put("Fiber", 4.0F);
                 return nutritionalValues;
             }
